fixed hotline styles
