+++ src/screens/Login/index.js
@@ -1,17 +1,13 @@
 import React, { useContext, useState, useEffect } from 'react';
 import { Context as AuthContext } from '../../state/AuthContext';
-import {
-  View,
-  Text,
-  TextInput,
-  StyleSheet,
-  TouchableOpacity,
-} from 'react-native';
+import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
 import { StyledView } from '../../styles/shared/StyledView';
 import {
   StyledButton,
   StyledButtonText,
 } from '../../styles/shared/StyledButton';
+import { StyledInputAuth } from '../../styles/shared/StyledInputAuth';
+
 import { useTranslation } from 'react-i18next';
 import AuthSVG from '_assets/svg/login.svg';
 
@@ -30,7 +26,6 @@ export default function LoginScreen({ navigation }) {
     const unsubscribe = navigation.addListener('blur', () => {
       removeErrors();
     });
-
     return unsubscribe;
   }, [navigation, removeErrors]);
 
@@ -38,16 +33,14 @@ export default function LoginScreen({ navigation }) {
     <StyledView style={styles.view}>
       <AuthSVG style={styles.background} />
       <Text style={styles.header}>{t('common.login')}</Text>
-      <TextInput
-        style={styles.input}
+      <StyledInputAuth
         placeholder="Email"
         autoCapitalize="none"
         placeholderTextColor="#6c757d"
         onChangeText={setEmail}
         value={email}
       />
-      <TextInput
-        style={styles.input}
+      <StyledInputAuth
         placeholder="Password"
         secureTextEntry={true}
         autoCapitalize="none"
@@ -62,7 +55,6 @@ export default function LoginScreen({ navigation }) {
         <TouchableOpacity onPress={() => navigation.navigate('Sign Up')}>
           <Text>Do not have an account? Go to sign up</Text>
         </TouchableOpacity>
-        <Text>or</Text>
         <TouchableOpacity onPress={() => navigation.navigate('Home')}>
           <Text>Proceed without login</Text>
         </TouchableOpacity>
@@ -75,17 +67,6 @@ export default function LoginScreen({ navigation }) {
 }
 
 const styles = StyleSheet.create({
-  input: {
-    width: 350,
-    height: 55,
-    backgroundColor: '#fff',
-    margin: 10,
-    padding: 8,
-    color: '#000',
-    borderRadius: 5,
-    fontSize: 18,
-    fontWeight: '500',
-  },
   view: {
     justifyContent: 'flex-end',
     height: '100%',
