+++ src/styles/colors.js
@@ -1,2 +1,3 @@
 export const primary = '#cadeee';
+export const secondary = '#FEF8E3';
 export const darkBlue = '#415889';
