+++ src/styles/shared/StyledInputAuth.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components/native'
+
+export const StyledInputAuth = styled.TextInput`
+    width: 350;
+    backgroundColor: #fff;
+    color: #000;
+    borderRadius: 5;
+    marginTop: 20;
+    fontSize: 18;
+    fontWeight: 400;
+    paddingLeft: 8;
+    paddingTop: 15;
+    paddingBottom: 15;
+`;
