+++ src/styles/shared/StyledView.js
@@ -3,5 +3,6 @@ import styled from 'styled-components/native'
 
 export const StyledView = styled.View`
   background-color: ${Colors.primary};
-  alignItems: center
+  flex: 1;
+  alignItems: center;
 `
